requirements.txt
manim
requests
streamlit
python-dotenv
google-generativeai
